N clusters medoids fix (#129)
* Don't sample with replacement for random placement

You'll always have an inertia at least as good as not doing this; i.e.
picking distinct medoids only helps in the optimization.

* Add test to ensure n_clusters

It appears to only be the "random" initialization that has the bug, but
it's cheap to test all initializations + methods.

* The initial medoids changed for determinism

We're now sampling differently so the inital deterministic sample is
different. This updates the test to use the new deterministic medoids.